% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8_1-correction-methods.R
\name{additive_UI}
\alias{additive_UI}
\alias{additive_server}
\alias{drift_server}
\alias{smooth_server}
\title{Apply an additive shift to data}
\usage{
additive_UI(id)

additive_server(id, sondeproj, y_var, index)

drift_server(id, sondeproj, y_var)

smooth_server(id, sondeproj, y_var, index)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{dates}{The date range to view the data.}

\item{period_view}{Should data be viewed by period?}

\item{p_length}{The length of the period to view.}
}
\value{
a reactive of length two with the min and max dates.
}
\description{
Used to adjust data either by a single value or apply a linear correction.
}
\keyword{internal}
