% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4-additive.R
\name{additive_UI}
\alias{additive_UI}
\alias{additive_server}
\title{Address any additive shifts}
\usage{
additive_UI(id)

additive_server(id, sdata, prj_path, log)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sdata}{A \code{reactiveVal} holding the current dataset.}

\item{prj_path}{A \code{reactiveVal} holding the path to save the project to.}

\item{log}{A \code{reactiveVal} holding the change log.}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset, user can select a group of points and apply a additive shift to the data to correct for shifts.
}
\keyword{internal}
