% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_1-logging-data-change.R
\name{apply_edit_UI}
\alias{apply_edit_UI}
\alias{apply_edit_server}
\title{Shiny module to added flags and save changes to  data}
\usage{
apply_edit_UI(id, edit_type = "flag", note = NULL, username)

apply_edit_server(id, sondeproj, edit, username, view_state)
}
\arguments{
\item{id}{the shiny ID of the action button}

\item{edit_type}{A character used to determine button naming options include flag, remove, interpolate, and change.}

\item{username}{A `reactiveVal` holding the name of the analyst for the changelog}

\item{sondeproj}{A `reactiveVal` holding the current `sondeproj`.}

\item{edit}{A `reactiveVal` holding a list of length six:
- data: new updated data as a `data.frame`
- rows: logical vector which specifies rows changed a TRUE
- y_var: parameter being edited
- step: name of the editing step for the changelog
- note: an optional note to add to the changelog
- flag: character flag to use for edits to the data}

\item{view_state}{A `reactiveVal` holding a list of items specifying the view state:
- abs_dates: The absolute range of dates within the dataset
- dates: The range of dates being viewed via the date selector
- period_view: Logical if the period view is being used
- period_length: Length of period view
- period_n: The period number to view.}
}
\description{
The UI creates a button to allow user to save the changes made to the data. The server function will provide a message to the user to
let them know the changes (or lack of changes made), add flags to the dataset, and save the dataset as a new version.
}
\keyword{internal}
