% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift-points.R
\name{apply_drift_shift}
\alias{apply_drift_shift}
\title{Apply a drift correction to a parameter}
\usage{
apply_drift_shift(x, rows, corrected, uncorrected)
}
\arguments{
\item{x}{Vector of data to apply correction to.}

\item{rows}{Row numbers of the values within x that should be corrected.}

\item{corrected}{Value of the corrected end value, used to determine how much to shift data.}

\item{uncorrected}{Value of the uncorrected value, used to determine how much to shift data.}
}
\value{
`x` with the drift shift applied.
}
\description{
Applies a linear correction to data based on the final "corrected" value often determined via a freshly calibrated sonde compared
to the sonde that's been deployed ("uncorrected").
}
\examples{
rows <- example_data$FileName == "example-data1.csv"
x_shift <- apply_drift_shift(example_data$fDOM_QSU, rows, 17.49, 21.71)
}
