% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-edit.R
\name{apply_edit}
\alias{apply_edit}
\title{Log edits to a `sondeproj`}
\usage{
apply_edit(proj, edit)
}
\arguments{
\item{proj}{A `sondeproj` object holding sonde data.}

\item{edit}{A list of length six:
- data: new updated data as a `data.frame`
- rows: logical vector which specifies rows changed as TRUE
- y_var: parameter being edited
- step: name of the editing step for the changelog
- note: an optional note to add to the changelog
- flag: character flag to use for edits to the data}
}
\value{
A `sondeproj` object with edits made.
}
\description{
Uses a list of edit parameters to update a `sondeproj` with the specified edits. Makes
changes to the data, changelog, and flags.
}
\examples{
data <- example_data
data$fDOM_QSU[1:4] <- NA
rows <- rep(FALSE, nrow(data))
rows[1:4] <- TRUE
edit <- list(data = example_data,
             rows = rows,
             y_var = "fDOM_QSU",
             step = "outlier removal",
             note = "example edit",
             flag = "RM07")
updated_proj <- apply_edit(example_sondeproj, edit)

}
