% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-interp.R
\name{apply_interp}
\alias{apply_interp}
\title{Map interpolated data back to dataset}
\usage{
apply_interp(data_fill, data_interp, y_var, max_length)
}
\arguments{
\item{data_fill}{\code{data.frame} based on \code{proj$data} with missing \code{datetime} values added from \code{prep_interp()}.}

\item{data_interp}{\code{data.frame} based on \code{proj$data} with duplicates condensed to a single value and missing values interpolated from \code{run_interp()}.}

\item{y_var}{Variable being interpolated.}

\item{max_length}{The maximum length in hours to fill via interpolation.}
}
\value{
\code{data_fill} with missing values interpolated.
}
\description{
Uses interpolated values to attempt to fill in missing data being aware of
maximum gap lengths to fill and duplicates.
}
\examples{
interp_df <- prep_interp(example_sondeproj)
filled_yvar <- run_interp(interp_df, "fDOM_QSU", "linear")
data_filled <- apply_interp(interp_df, filled_yvar,
                            "fDOM_QSU", 8)
}
