% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagging-module.R
\name{confirm_changes_UI}
\alias{confirm_changes_UI}
\alias{confirm_changes_server}
\title{Shiny module to added flags and save changes to  data}
\usage{
confirm_changes_UI(id, note = NULL)

confirm_changes_server(
  id,
  newdata,
  sdata,
  index = NULL,
  par,
  flag_name,
  note,
  prj_path,
  log
)
}
\arguments{
\item{id}{the shiny ID of the action button}

\item{note}{a note from the analyst about the change made}

\item{newdata}{the sonde data with the appropriate transformation performed relative to the flags being added}

\item{sdata}{the sonde data.frame that will be flagged (reactive)}

\item{index}{the index values of the rows to flag in data}

\item{par}{the parameter to flag}

\item{flag_name}{a character with the name of the flag}

\item{prj_path}{the file path to save the sonde project to (reactive)}

\item{log}{A `reactiveVal` holding the change log.}
}
\description{
The UI creates a button to allow user to save the changes made to the data. The server function will provide a message to the user to
let them know the changes (or lack of changes made), add flags to the dataset, and save the dataset as a new version.
}
\keyword{internal}
