% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdom-correction.R
\name{correct_fdom}
\alias{correct_fdom}
\title{Apply fDOM temperature and turbidity corrections}
\usage{
correct_fdom(proj, temp = NULL, turb = NULL)
}
\arguments{
\item{proj}{A \code{sondeproj} object.}

\item{temp}{Either \code{NULL} or a list object (see details) for temperature correction.}

\item{turb}{Either \code{NULL} or a list object (see details) for turbidity correction.}
}
\value{
A \code{data.frame} object with fDOM updated with the corrections.
}
\description{
Takes functions and coefficient values and applies them to fDOM data within a sonde project to apply fDOM corrections.
Uses flags stored within the project to only correct data that hasn't previously been corrected. Additionally,
turbidity corrections will only be applied to temperature corrected data.
}
\details{
To apply a correction you must supply a list object with the following structure:
\itemize{
\item params: a named list with parameter values where the names match the arguments within the function
\item fun: a function to apply to the fDOM data where the first argument is fDOM, the second is either
temperature or turbidity, and the third is parameters within the function, pulled from \code{params}.
}
}
\examples{
temp <- list(params = list(rho = -0.011),
             fun = function(fdom, temp, parms){fdom / (1 + parms$rho*(temp - 25))})
corr_data <- correct_fdom(example_sondeproj, temp=temp, turb=NULL)

}
