% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_5-data-version.R
\name{data_version_UI}
\alias{data_version_UI}
\alias{data_version_server}
\title{Keep track of visualized version and edits}
\usage{
data_version_UI(id)

data_version_server(id, sondeproj, y_var, row)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{y_var}{A \code{reactiveVal} holding the Y-variable to plot on the y-axis.}

\item{row}{A \code{reactiveVal} holding the number of the row selected in the changelog.}
}
\value{
A list containing:
\itemize{
\item \code{current}: Reactive expression containing the version of the data corresponding to the selected changelog row.
\item \code{changed}: Reactive expression containing the previous values for #' the selected parameter when changes are being displayed.
}
}
\description{
Keeps track of the version selected and returns the correct data to plot with any edits to view.
}
\keyword{internal}
