% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{write_data}
\alias{write_data}
\alias{get_data}
\alias{clear_data}
\alias{set_data}
\title{Sonde data versioning}
\usage{
write_data(data, version, env = .pkgenv)

get_data(env = .pkgenv)

clear_data(env = .pkgenv)

set_data(data_ver, env = .pkgenv)
}
\arguments{
\item{data}{a data.frame to add to the version history}

\item{version}{the name of the version}

\item{env}{the environment in which configuration settings is stored. Defaults to the package environment.}

\item{data_ver}{a existing data version history read in from read_project}
}
\value{
a list of \code{data.frames}
}
\description{
Stores edits to the raw sonde data as objects in a list so edits can be undone.
}
\examples{
get_data()

write_data(raw_sonde, "V1")
}
