% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-interp.R
\name{data_interp}
\alias{data_interp}
\title{Fill data gaps with interpolation}
\usage{
data_interp(proj, y_var, method, max_length, freq = 1, progress = NULL)
}
\arguments{
\item{proj}{A \code{sondeproj} object holding sonde data.}

\item{y_var}{Variable being interpolated.}

\item{method}{The method to use for interpolation, options include: \code{linear}, \code{spline}, \code{random_forest}, and \code{ts_interp} see details for further information about the methods.}

\item{max_length}{The maximum length in hours to fill via interpolation.}

\item{freq}{The period interval for the time series in days, only used if \code{method} is \code{ts_interp}.}

\item{progress}{A function passed from a shiny module to create a progress bar.}
}
\value{
A data.frame if there were completely missing observations, rows may have been added with the missing datetimes. Additionally, an extra column is added (\code{x_fill})
which is a logical flag identifying which values were interpolated.
}
\description{
Provides a number of methods to fill NA values for a given parameter within the dataset with
options to only fill a certain gap length otherwise leave the missing values alone.
}
\details{
Currently the follow interpolation options are supported:
\itemize{
\item \code{linear}: Uses the zoo::na.approx() function to linearly fill gaps.
\item \code{spline}: Uses the zoo::na.spline() function to fill gaps via spline interpolation (makes smoother curves).
\item \code{random_forest}: Depending on how much data is in the project, this option can be a little slow. Uses missForest::missForest() to
fill gaps. uses responses accross all the available parameters to try and fill in the missing parameter.
\item \code{ts_interp}: Uses forecast::na.interp() which treats the data as a time series and make predictions based on the natural seasonal
patterns within the data. The seasonality is determined by \code{freq}. For instance, if set to 1, it will account for daily fluctuations.
If set to 365 it wil look at annual fluctuations.
}
}
\examples{
path <- file.path(fs::path_package("extdata", package = "SondePolishR"),
"example-sondeproj-messy.RDS")
proj <- readRDS(path)
data_filled <- data_interp(proj, "fDOM_QSU", 4, method="linear")

}
