% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-data.R
\name{describe_data}
\alias{describe_data}
\title{Gets summary statistics for Sonde data}
\usage{
describe_data(data, precip = NULL)
}
\arguments{
\item{data}{a \code{data.frame} with sonde dataset.}

\item{precip}{Optional precipitation dataset.}
}
\value{
a \code{data.frame}
}
\description{
Used to quickly generate summary statistics about each parameter for a given dataset.
}
\examples{
describe_data(example_data)
}
