% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8_1-correction-methods.R
\name{drift_UI}
\alias{drift_UI}
\title{Apply a drift correction to a data file}
\usage{
drift_UI(id, sondeproj)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{dates}{The date range to view the data.}

\item{period_view}{Should data be viewed by period?}

\item{p_length}{The length of the period to view.}
}
\value{
a reactive of length two with the min and max dates.
}
\description{
Used to account for instrument drift by applying a linear correction to a data file. Uses differences between
check and resident sonde measurments when available as the default correction amount.
}
\keyword{internal}
