% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{US EPA Level III Ecoregions}
\format{
An sf object with 87 rows and 9 columns:
\itemize{
\item \strong{US_L3CODE}: Code for Level III Ecoregion (US)
\item \strong{US_L3NAME}: Name for Level III Ecoregion (US)
\item \strong{NA_L3CODE}: Code for Level III Ecoregion (North America/CEC)
\item \strong{NA_L3NAME}: Name for Level III Ecoregion (North America/CEC)
\item \strong{NA_L2CODE}: Code for Level II Ecoregion (North America/CEC)
\item \strong{NA_L2NAME}: Name for Level II Ecoregion (North America/CEC)
\item \strong{NA_L1CODE}: Code for Level I Ecoregion (North America/CEC)
\item \strong{NA_L1NAME}: Name for Level I Ecoregion (North America/CEC)
\item \strong{geometry}: Spatial data associated with the polygon
}
}
\source{
US Environmental Protection Agency. (2013). Level III Ecoregions of the Conterminous United States \href{ftp://ftp.epa.gov/wed/ecoregions/us/us_eco_l3.zip}{Data set}.
U.S. EPA Office of Research and Development (ORD) - National Health and Environmental Effects Research Laboratory (NHEERL).
}
\usage{
ecoregions
}
\description{
Designed to serve as a spatial framework for the research, assessment, and
monitoring of ecosystems and ecosystem components, ecoregions denote areas of
similarity in the mosaic of biotic, abiotic, terrestrial, and aquatic ecosystem
components with humans being considered as part of the biota.
}
\note{
Data has been simplified from the original dataset to reduce data size using the
sf::st_simplify() function.
}
\keyword{datasets}
