% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_calcheck}
\alias{example_calcheck}
\title{Sample Calibration Check Data}
\format{
A \code{data.frame} with the following columns:
\itemize{
\item \strong{Date}: The date of the site visit.
\item \strong{Site_Code}: The site name or site code.
\item \strong{Parameter}: The name of the parameter, should match column names from read_sonde.
\item \strong{Resident_Probe_Serial}: The serial number for the sonde located at the site.
\item \strong{Resident_Value}: The measured value for the parameter on the resident sonde.
\item \strong{Check_Probe_Serial}: The serial number for the sonde used to check the calibration. Should be freshly calibrated.
\item \strong{Check_Value}: The measured value for the parameter on the check sonde.
\item \strong{Notes}: Notes associated with the calibration check or data entry.
}
}
\usage{
example_calcheck
}
\description{
An example of a calibration check form which tracks the results of calibration
checks during field visits. This data is collected by comparing measurements of the
"resident" sonde directly next to a freshly calibrated "check" sonde.
}
\keyword{datasets}
