% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_ver}
\alias{example_data_ver}
\title{Sample Data Versioning}
\format{
A list where the names indicate the version name and each list item is a version of the dataset. The
first item is special:
\itemize{
\item \strong{raw}: the sonde data before any changes have been made
}
}
\usage{
example_data_ver
}
\description{
An example of the data versioning object saved to the package environment as \code{data_ver}.
}
\keyword{datasets}
