% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_fieldform}
\alias{example_fieldform}
\title{Sample Field Form}
\format{
A \code{data.frame} with the following columns:
\itemize{
\item \strong{Date}: The date of the site visit.
\item \strong{Site_Code}: The site name or site code.
\item \strong{Time}: The start time of the field visit (used to guess out of water periods when data is missing).
\item \strong{Start_Sonde_Serial}: The serial number for the sonde in the water at the start of the visit.
\item \strong{Start_Sonde_Name}: The sonde name for the sonde in the water at the start of the visit.
\item \strong{End_Sonde_Serial}: The serial number for the sonde in the water at the end of the visit.
\item \strong{End_Sonde_Name}: The sonde name for the sonde in the water at the end of the visit.
\item \strong{Removal_Time}: The time the sonde was removed from the water.
\item \strong{Return_Time}: The time the sonde was returned to the water.
\item \strong{Next_Timepoint}: The next timepoint that data will be collected.
This should be the next "good" time point after the sonde has been returned to the water, but is sometimes the timepoint
when the wiper is checked and the sonde is out of the water.
\item \strong{Data_Download}: Logical, was data downloaded at this visit?
\item \strong{Download_Device}: The name of the device data was downloaded to.
\item \strong{Remove_Period}: Logical, is there a data disruption that merits removing data during the out of water period. Used to skip periods
where we have coarse out of water periods (missing times), but the data appears uninterrupted to prevent removing excess data.
\item \strong{Crew}: The names or initials of the people performing the field visit.
\item \strong{Weather}: A description of the weather during the field visit.
\item \strong{Notes}: Notes associated with the field visit.
\item \strong{DataEntry_Notes}: Additional notes from data entry. Can have any name that includes "Notes"
}
}
\usage{
example_fieldform
}
\description{
An example of a field form used to collect information about a field visit to
check and perform maintenance. Critically this information is used to track periods
when the sonde was out of the water.
}
\keyword{datasets}
