% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_log}
\alias{example_log}
\title{Sample Data Change Log}
\format{
An data.frame object with 2 rows and 6 columns:
\itemize{
\item \strong{datetime}: The date and time the change was made
\item \strong{parameter}: The parameter that was changed
\item \strong{step}: The name of the correction step performed
\item \strong{n_changed}: The number of data points modified
\item \strong{user}: The name of the person who made the change
\item \strong{version}: the version name indicated by a hash code
}
}
\usage{
example_log
}
\description{
An example of the change log object saved to the package environment as \code{log}. This tracks the changes made to the dataset.
}
\keyword{datasets}
