% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_precip}
\alias{example_precip}
\title{Sample Precipitation Data}
\format{
A \code{data.frame} with the following columns:
\itemize{
\item \strong{DateTime}: The date and time of the observation.
\item \strong{Precip_mm_hr}: The hourly precipitation value for that hour in mm.
}
}
\source{
The average MERRA-2 bias corrected total precipitation at the surface of the earth. NASA POWER Project.
NASA Langley Research Center (LaRC), Prediction Of Worldwide Energy Resources (POWER),
accessed 2026-06-26. https://power.larc.nasa.gov/parameters/?parameter=PRECTOTCORR
}
\usage{
example_precip
}
\description{
An example of a precipitation data file used to plot with sonde data to help determine if
peak values are real (particularly for fDOM and turbidity).
}
\keyword{datasets}
