% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_project}
\alias{example_project}
\title{Sample Sonde Project}
\format{
A list where the names indicate the version name and each list item is a version of the dataset. The
first two items are special:
\itemize{
\item \strong{log}: the change log
\item \strong{raw}: the sonde data before any changes have been made
}
}
\usage{
example_project
}
\description{
An example of a sonde project, which saves the versions and changes made to the raw data.
}
\keyword{datasets}
