% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sondeproj}
\alias{example_sondeproj}
\title{Sample Sonde Project}
\format{
An object of class \code{sondeproj} with the following elements:
\itemize{
\item \strong{meta}: A list of length three:
\item \strong{site}: A character with the sonde site name.
\item \strong{tz}: A character with the site timezone.
\item \strong{coords}: A vector of length two with the latitude and longitude of the site.
\item \strong{data}: A \code{data.frame} of the dataset, updated with any changes.
\item \strong{precip}: A \code{data.frame} with  precipitation values.
\item \strong{fieldform}: A \code{data.frame} with the field form data. See \code{example_fieldform} for details on the structure.
\item \strong{calcheck}: A \code{data.frame} with the calibration check data. See \code{example_calcheck} for details on the structure.
\item \strong{diffs}: A list of \code{data_diff} which stores the changes made to \code{data}. See \code{daff} package for more details on this structure.
\item \strong{changelog}: A \code{data.frame} which stores a summary of the changes made in each \code{data_diff}:
\item \strong{datetime}: The date and time the change was made.
\item \strong{parameter}:  The parameter that the change was made to.
\item \strong{step}: The name of the processing step the change was made in.
\item \strong{nchanged}: The number of points that were changed.
\item \strong{note}: Notes added about the changes made.
\item \strong{user}: The username of the person who made the change.
\item \strong{diff_name}: The name of the associated \code{data.diff}.
}
}
\usage{
example_sondeproj
}
\description{
An example of a sonde project, which is a list type object that stores the data, associated metadata, and
tracks changes made to the data.
}
\keyword{datasets}
