% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2-explore-plot.R
\name{explore_data_UI}
\alias{explore_data_UI}
\alias{explore_data_server}
\title{Visualize dataset by analyte}
\usage{
explore_data_UI(id)

explore_data_server(
  id,
  sondeproj,
  data_ver,
  y_var,
  view_state,
  username,
  webgl_supported
)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{view_state}{A \code{reactiveVal} holding a list of items specifying the view state:
\itemize{
\item abs_dates: The absolute range of dates within the dataset
\item dates: The range of dates being viewed via the date selector
\item period_view: Logical if the period view is being used
\item period_length: Length of period view
\item period_n: The period number to view.
}}

\item{username}{A \code{reactiveVal} holding the name of the analyst for the changelog}

\item{webgl_supported}{A \code{reactiveVal} indicating if webgl is support in the current browser.}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset with options to explore full dataset or view the data in weekly sections.
Plots are created using \code{plotly} so they are interactive. Module also allows the user to view
the changes over the dataset versions via row selection in a table via the \code{log}.
}
\keyword{internal}
