% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step10_1-exporting-data.R
\name{save_path_UI}
\alias{save_path_UI}
\alias{save_path_server}
\title{File save selector UI}
\usage{
save_path_UI(id, button_label = "Export")

save_path_server(
  id,
  data,
  startname = "sonde_export",
  label = "Choose Location",
  title = "Select save path",
  filetype = reactiveVal(".csv"),
  username,
  type,
  data_ver
)
}
\arguments{
\item{id}{module id}

\item{data}{data to save to specified file on click}

\item{startname}{the default name for the file}

\item{label}{button label}

\item{title}{dialog title}

\item{filetype}{file extension}

\item{username}{A \code{reactiveVal} holding the name of the analyst for the changelog}

\item{type}{a \code{reactiveVal} holding a character specifying the type of data being saved options include
"sondeproj", "data", "dups", "gaps", "changelog", "precip", "calcheck", "fieldform", "change-report"}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}
}
\description{
File save selector UI
}
\keyword{internal}
