% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{flag_data}
\alias{flag_data}
\title{Flag dataset and save a copy}
\usage{
flag_data(
  data,
  par,
  flag_name,
  index,
  note = "",
  prj_path = NULL,
  makeNA = FALSE
)
}
\arguments{
\item{data}{a data.frame with sonde data}

\item{par}{the parameter to check}

\item{flag_name}{a character with the name of the flag}

\item{index}{the index values for the rows to be flagged}

\item{note}{a note from the analyst about the change made}

\item{prj_path}{the file path to save the sonde project to, if \code{NULL} will try to use \code{prj_path} stored in package environment.}

\item{makeNA}{if \code{TRUE} the flagged data will be converted to NA values}
}
\value{
returns the flagged copy of the \code{data.frame}
}
\description{
Adds flags to the dataset, saves the new version of the dataset, and saves the project to the
project path specified by the user as a \code{.RDS} object. Also provides messages about the
changes being made.
}
\examples{
prj_path <- file.path(withr::local_tempdir(), "test_prj.RDS")

#do intial versioning things (automatic in app)
  clear_data()
  clear_log()
  write_data(raw_sonde, "raw")

#flag data
  data <- flag_data(raw_sonde, "fDOM_QSU", "test_flag", 1:4, "testing flags", prj_path)
}
