% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a change report from a sonde project}
\usage{
generate_report(proj, username, export_path)
}
\arguments{
\item{proj}{the \code{sondeproj} you want to generate the report for}

\item{username}{the username of the person generating the report}

\item{export_path}{the path including the file name to save the report to, should end with a \code{.pdf}}
}
\value{
a \code{.pdf} document with a page for each line in the change log of the \code{sondeproj}.
}
\description{
Use the version control information to create a \code{.pdf} report
}
\examples{
temp_report <- withr::local_tempfile(fileext = ".pdf")
generate_report(example_sondeproj, "Smith", temp_report)
}
