% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{get_diff}
\alias{get_diff}
\title{Track differences between sonde data}
\usage{
get_diff(olddata, newdata, id = c("DateTime_rd", "DupNum"), ignore = "Index")
}
\arguments{
\item{olddata}{a \code{data.frame} with the original data.}

\item{newdata}{a \code{data.frame} with the revised data.}

\item{id}{name of the column name used to match observations between \code{olddata} and \code{newdata}. Must be convertible to a number.}

\item{ignore}{a character vector including any column names to not track.}
}
\value{
A \code{diff} object with a named list item for each column being tracked.
Each list item will either be \code{NULL} if there were no changes to that column or have the following structure:
\itemize{
\item op_type: a character describing the type of change made
\item id: the id values for the change made
\item new_data: the values of the changed values in \code{newdata}
\item old_data: the values of the changed values in \code{olddata}
}
}
\description{
Compares the values between the two \code{data.frames} and creates a small \code{diff} object that stores the
datetimes of the change and the new and old values. Used to move between versions. Note that this function
does not currently support addition and removal of columns, but row additions/subtractions are supported.
}
\examples{
data1 <- example_data[1:10,]
data2 <- data1
data2$fDOM_QSU[1:4] <- NA
get_diff(data1, data2)
}
