% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ecoregion.R
\name{get_ecoregion}
\alias{get_ecoregion}
\title{Get the US EPA Ecoregion at a Site}
\usage{
get_ecoregion(site = NULL, geometry = TRUE, lvl = 3)
}
\arguments{
\item{site}{a character vector in the form lat, long or sf object of length 1}

\item{geometry}{logical, should geometry be returned?}

\item{lvl}{a number (1,2, or 3) indicating the EPA ecoregion level to return (only used if geometry = FALSE)}
}
\value{
if geometry is TRUE
 - a sf object of the ecoregion the site is in
if geometry is FALSE
 - a character vector returing the name of the ecoregion
}
\description{
Identities the US EPA Ecoregion at a provided latitude and longitude.
}
\examples{
get_ecoregion(c("44.20",	"-122.2"))
}
