% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_encoding}
\alias{get_encoding}
\title{Try to guess csv file encoding}
\usage{
get_encoding(file)
}
\arguments{
\item{file}{file path to sonde data}
}
\value{
A character with the file encoding
}
\description{
Turns out it's actually quite challenging to load in \code{.csv} file when it has a weird encoding. This will use \link[readr]{guess_encoding}
first to see if it can figure it out, if not, it will test the two likely encoding: UTF-16LE or Windows-1252.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-example.csv")
get_encoding(file)
}
