% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-precip.R
\name{get_precip}
\alias{get_precip}
\title{Get hourly precipitation at data site}
\usage{
get_precip(data, lat, long, method, token = NULL)
}
\arguments{
\item{data}{the data to get matching precipitation data for}

\item{lat}{the latitude to get precipitation data at}

\item{long}{the longitude to get precipitation data at}

\item{method}{method used to get data either "merra-2" or "nldas"}

\item{token}{only required for nldas method. see details for how to obtain this token.}
}
\value{
a data.frame with two columns:
\itemize{
\item DateTime: The datetime (\code{POSIXct}) in the same timezone as the data, at an hourly resolution.
\item Precip_mm_hr: Average precipitation at the requested point in mm per hour.
}
}
\description{
Precipitation is downloaded at an hourly scale based on the provided
coordinates.
}
\details{
There are two different datasets that can be downloaded via this function:
\itemize{
\item \href{https://gmao.gsfc.nasa.gov/gmao-products/merra-2/}{Merra-2}: Available from NASA power. This dataset is available
across a global scale at a resolution of 0.5 x 0.625 degrees available from 1981 to near real time
\item \href{https://ldas.gsfc.nasa.gov/nldas}{NLDAS}: Available from NASA Earthdata. This dataset is available
across CONUS at a resolution of 0.125 × 0.125 degrees available from 1981 to near real time. This data requires a
token to access the data. See \href{https://urs.earthdata.nasa.gov/documentation/for_users/user_token}{here} for
directions on creating a token. Note that this token should be kept secret. You will also need to authorize the NASA GESDISC DATA ARCHIVE
application. For more detailed directions see \code{vignette("using-precipitation-data")}
}
}
\examples{
data <- example_data[example_data$Date == "2024-11-13",]
precip <- get_precip(data, 43.96775, -122.63012, "merra-2")
}
