% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_skip}
\alias{get_skip}
\title{Guess the number of rows to skip to get correct headers}
\usage{
get_skip(file, encoding = NULL)
}
\arguments{
\item{file}{file path to sonde data (.xlsx or .csv)}

\item{encoding}{the file encoding if file is a .csv, will guess if NULL}
}
\value{
a numeric indicating the number of rows to skip
}
\description{
Guess the number of rows to skip to get correct headers
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "example-csv-data1.csv")
get_skip(file)

file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-usb-example.csv")
get_skip(file)
}
