% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_dups.R
\name{identify_dups}
\alias{identify_dups}
\title{Identify duplicated observations with sonde data}
\usage{
identify_dups(data)
}
\arguments{
\item{data}{a \code{data.frame} with Sonde data.}
}
\value{
a \code{data.frame} with the following columns with a row for each duplicated period:
\itemize{
\item start: starting datetime of the duplicated data
\item end: ending datetime of the duplicated data
\item duptype: either "same file" or "multiple files" indicating the type of duplication
\item ndif: number of points with the duplicated section that are not exactly the same
\item length: number of points in the duplicated section
\item perc_dif: percentage of the data that's different
\item likely_issue: based on information about the duplicates, a guess about the cause of the duplicate
\item user_note: user input information about the duplicated section
}
}
\description{
Uses the interval-rounded datetime (\code{DateTime_rd}) observations to identify periods of duplicated
observations. Based on if they match and the length of the section it identifies a likely cause of
the duplication.
}
\examples{
identify_dups(example_data)
}
