% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-project.R
\name{load_project}
\alias{load_project}
\title{Load and combine sonde project components}
\usage{
load_project(
  csv_path = NULL,
  csv_files = NULL,
  prj_path = NULL,
  ff_path = NULL,
  cc_path = NULL,
  tz = "Etc/GMT+8",
  site = NULL,
  username = NULL,
  update_pb = NULL
)
}
\arguments{
\item{csv_path}{character vector with the filepaths to sonde data files,
can be more than one.}

\item{csv_files}{names of the csv files, primarily used for shiny input which
doesn't store filenames with paths.}

\item{prj_path}{character vector with the filepath to a sonde project.}

\item{ff_path}{character vector with the filepath to a field form file.}

\item{cc_path}{character vector with the filepath to a calibration check file.}

\item{tz}{the timezone for the sonde data}

\item{site}{the site name or site code.}

\item{username}{the username of the person who made the change}

\item{update_pb}{takes a function used to update a progress bar in a shiny
interface.}
}
\value{
a \code{sondeproj} object. For more details on structure
see \code{example_sondeproj}
}
\description{
Uses provided filepaths and to read in, clean/format, and merge together components of a \code{sondeproj}.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "example-csv-data1.csv")
proj <- load_project(csv_path = file, csv_files = "example_file1", username="Smith")
}
