% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj-tracking.R
\name{new_version}
\alias{new_version}
\title{Checks if the data is different than the current saved version}
\usage{
new_version(data)
}
\arguments{
\item{data}{the data.frame to check to see if it's new}
}
\value{
TRUE if data differs from the previous version,
FALSE if data is the same as the previous version
}
\description{
Checks if the data is different than the current saved version
}
