% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nice_tz}
\alias{nice_tz}
\title{Make nice timezones for selecting}
\usage{
nice_tz()
}
\value{
a vector of time zones with names that are human readable
}
\description{
Uses \link[base]{OlsonNames} to extract potential timezones for the .csv file.
}
