% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nice_yvar}
\alias{nice_yvar}
\title{Reformats column names to be human readable}
\usage{
nice_yvar(df)
}
\arguments{
\item{df}{the dataframe you want to column names from}
}
\value{
a named vector where the names are the human readable names and the values are the column names
}
\description{
Reformats column names to be human readable
}
