% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phys_limits}
\alias{phys_limits}
\title{Physical Limits for Sonde Parameters}
\format{
An data.frame object with 29 rows and 5 columns:
\itemize{
\item \strong{parameter}: USGS parameter code
\item \strong{ecoregion}: Name for Level III Ecoregion (US)
\item \strong{max}: Code for Level III Ecoregion (North America/CEC)
\item \strong{min}: Name for Level III Ecoregion (North America/CEC)
\item \strong{metric}: Sonde column name for parameter
}
}
\usage{
phys_limits
}
\description{
Physical minimum and maximum limits used to perform preliminary filtering of sonde data.
Determined by ecoregion using USGS water quality data within each ecoregion.
}
\details{
\itemize{
\item Minimum was determined as the minimum daily value observed across any site within the ecoregion.
\item Maximum was determined as the highest value between the maximum of the daily minimum or the 99.9th percentile of the daily maximums.
}

Only accepted data values were used in the calculations. See DATASET.R for exact process used to derived minimums and maximums.
}
\keyword{datasets}
