% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3-physical-lim.R
\name{phys_limits_UI}
\alias{phys_limits_UI}
\alias{phys_limits_server}
\title{Flagging data the is outside physical limits}
\usage{
phys_limits_UI(id)

phys_limits_server(id, data, prj_path)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{data}{A reactive holding the loaded dataset.}

\item{prj_path}{remove in function}
}
\description{
There are certain thresholds for some of the sonde parameters that aren't physical possible (i.e, water temperature above 100 deg C).
This module visualizes those limits and flags data outside the limits.
}
\keyword{interal}
