% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical-limit.R
\name{physical_limit}
\alias{physical_limit}
\title{Remove data outside of physical limits}
\usage{
physical_limit(df, min, max, par, keep = NULL)
}
\arguments{
\item{df}{a data.frame with sonde data}

\item{min}{the minimum physical value}

\item{max}{the maximum physical value}

\item{par}{the parameter to check}

\item{keep}{row numbers to exclude from the check (used to omit values which user has said are acceptable)}
}
\value{
a list with two data.frames:
\itemize{
\item \strong{outlier:} the rows of df that have values outside the limits not excluded using keep
\item \strong{within:} the rows of df that have values within the limits or have been kept using \code{keep}
}
}
\description{
A good first QA/QC check for sonde data is to determine if there are any values that aren't physically possible.
}
\examples{
results <- physical_limit(raw_sonde, min=0, max=98, par="ODO_sat")
rownames(results$outlier)

results <- physical_limit(raw_sonde, min=0, max=98, par="ODO_sat", keep=1436)
rownames(results$outlier)

}
