% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-plotting-module.R
\name{plotting_data_UI}
\alias{plotting_data_UI}
\alias{plotting_data_server}
\title{Shiny module to plot data from sondeproj with options}
\usage{
plotting_data_UI(id)

plotting_data_server(id, sondeproj, choices_fun = NULL)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}
}
\value{
a \code{plotly} objection
}
\description{
Used to nicely package all the plotting options so it can be reused.
}
\keyword{internal}
