% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{plot_sonde}
\alias{plot_sonde}
\title{Create plotly object of requested sonde data}
\usage{
plot_sonde(
  data,
  y_var,
  y2_var = NULL,
  proj = NULL,
  opts = list(points = TRUE, line = TRUE, files = FALSE, oow = FALSE, calcheck = FALSE,
    qualflag = FALSE),
  source = "plot"
)
}
\arguments{
\item{data}{A \code{data.frame} with the sonde data to plot.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{y2_var}{A second, optional y-variable to plot on a second axis.}

\item{proj}{A \code{sondeproj} object with additional metadata to plot,
only required if opts includes oow, calcheck, quality_flag or \code{y2_var} is "precip" or "raw".}

\item{opts}{A list of options for plotting:
\itemize{
\item points: should points be plotted?
\item line: should line be plotted?
-files: should points be colored by file?
-oow: should out of water periods be plotted?
-calcheck: should cal check be plotted?
-quality_flag: should questionable points be plotted?
}}

\item{source}{A character specifying the plot name for shiny reactivity.}
}
\value{
a \code{plotly} object
}
\description{
Allows options for iterative plotting sonde project data for different variables,
used within modules to simplify code for plotting.
}
\examples{
plot_sonde(example_data, y_var = "Temp_C")
plot_sonde(example_data, y_var = "Temp_C", y2_var = "ODO_mg_L") #adding a second axis
plot_sonde(example_data, proj = example_sondeproj,
y_var ="Temp_C", y2_var = "precip")
}
