% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-func.R
\name{preserve_zoom}
\alias{preserve_zoom}
\title{Store user zoom info from plotly to maintain zoom while updating data}
\usage{
preserve_zoom(df, y_var, plot_id)
}
\arguments{
\item{df}{the data.frame with the sonde data}

\item{y_var}{the y_var}

\item{plot_id}{the plot_id of the plot to keep zoom steady}
}
\value{
a list of length two:
\itemize{
\item xaxis$range with the start and end values for the x-axis range
\item yaxis$range with the start and end values for the y-axis range
}
}
\description{
Natively, when plotly redraws, it will remove any user interactions like zooming. This function
will capture the zoom level and apply it when redrawing.
}
