% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj-tracking.R
\name{get_prjpath}
\alias{get_prjpath}
\alias{set_prjpath}
\alias{clear_prjpath}
\title{Get and set project path variable in package environment}
\usage{
get_prjpath(env = .pkgenv)

set_prjpath(prj_path, env = .pkgenv)

clear_prjpath(prj_path, env = .pkgenv)
}
\arguments{
\item{env}{the environment in which configuration settings is stored. Defaults to the package environment.}

\item{prj_path}{the file path including the file name to the save location of the Sonde project.}
}
\description{
Used to store the save path to a sonde project so the user doesn't
need to specify each time.
}
