% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4-quality-flags.R
\name{quality_UI}
\alias{quality_UI}
\alias{quality_server}
\title{Flag data as questionable}
\usage{
quality_UI(id)

quality_server(id, sondeproj, data_ver, y_var, view_state)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{view_state}{A `reactiveVal` holding a list of items specifying the view state:
- abs_dates: The absolute range of dates within the dataset
- dates: The range of dates being viewed via the date selector
- period_view: Logical if the period view is being used
- period_length: Length of period view
- period_n: The period number to view.}
}
\description{
Allows the user to mark data that looks for "weird" and can be viewed as a plotting option or
used to auto-select as an outlier.
}
\keyword{internal}
