% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_sonde}
\alias{raw_sonde}
\title{Sample Sonde Data}
\format{
An data.frame object with 2209 rows and 23 columns:
\itemize{
\item \strong{Date_MM_DD_YYYY}: Date of measurement
\item \strong{Time_HH_mm_ss}: Time of measurement
\item \strong{DateTime}: Date and time of measurement as a POSIXct
\item \strong{Time_Fract_Sec}: Time in fractions of a second (usually not used)
\item \strong{Site_Name}: Name of the site
-\strong{Cond_/u00B5S_cm} through \strong{Cable_Pwr_V}: Sonde measurement variables
}
}
\source{
Forest Ecohydrology and Watershed Science Lab (2024) Fall Creek: 2024-04-22 to 2024-05-15. Data set.
}
\usage{
raw_sonde
}
\description{
An example of the output from \link[SondePolishR]{read_sonde} function.
}
\keyword{datasets}
