% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonde-project.R
\name{save_project}
\alias{save_project}
\alias{read_project}
\title{Save version history and sonde data as a project}
\usage{
save_project(data_ver, log, prj_path)

read_project(prj_path)
}
\arguments{
\item{data_ver}{the data version object (a list)}

\item{log}{the change log object (a data.frame)}

\item{prj_path}{the path to save the project to including the file name}
}
\value{
\code{save_project} saves \code{prj_path}, \code{data_ver} and \code{log} as a \code{.RDS} file to the provided \code{prj_path}.
The format of the \code{.RDS} object is a list with the following items:
\itemize{
\item \strong{prj_path:} a character vector specifying the save location, this is used to guess the save location when reading in a project.
\item \strong{change_log:} a \code{data.frame} object detailing the changes made to the dataset.
\item \strong{raw:} a \code{data.frame} with the initial dataset as it was loaded in.
\item \strong{...:} other versions of the data specified by hash codes.
}

\code{read_project} loads the project, updates the package environment with the project data, and returns \code{data_ver} as a list of \code{data.frames}
}
\description{
Combines the change history with the version history of the sonde data and saves to a
\code{.RDS} file to prevent data loss and enable work over multiple days.
}
