% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-meta.R
\name{read_cal}
\alias{read_cal}
\title{Read sonde calibration check file}
\usage{
read_cal(file, tz)
}
\arguments{
\item{file}{File path to calibration check data (.csv).}

\item{tz}{Timezone for the times in the form.}
}
\value{
a \code{data.frame} with the following columns. The \code{file} to read in should have the same structure:
\itemize{
\item \strong{Date}: The date of the site visit.
\item \strong{Site_Code}: The site name or site code.
\item \strong{Parameter}: The name of the parameter, should match column names from read_sonde.
\item \strong{Resident_Probe_Serial}: The serial number for the sonde located at the site.
\item \strong{Resident_Value}: The measured value for the parameter on the resident sonde.
\item \strong{Check_Probe_Serial}: The serial number for the sonde used to check the calibration. Should be freshly calibrated.
\item \strong{Check_Value}: The measured value for the parameter on the check sonde.
\item \strong{Notes}: Notes associated with the calibration check or data entry.
}
}
\description{
Loads a \code{.csv} file with calibration checks performed during field visits.
Ensures correct column classes and converts NA to R recognized NA values.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "example-calcheck.csv")
calcheck <- read_cal(file, tz="Etc/GMT+8")
}
