% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-meta.R
\name{read_ff}
\alias{read_ff}
\title{Read sonde field form}
\usage{
read_ff(file, tz)
}
\arguments{
\item{file}{File path to field form data (.csv).}

\item{tz}{Timezone for the times in the form.}
}
\value{
a \code{data.frame} with the following columns. The \code{file} to read in should have the same structure:
\itemize{
\item \strong{Date}: The date of the site visit.
\item \strong{Site_Code}: The site name or site code.
\item \strong{Time}: The start time of the field visit (used to guess out of water periods when data is missing).
\item \strong{Start_Sonde_Serial}: The serial number for the sonde in the water at the start of the visit.
\item \strong{Start_Sonde_Name}: The sonde name for the sonde in the water at the start of the visit.
\item \strong{End_Sonde_Serial}: The serial number for the sonde in the water at the end of the visit.
\item \strong{End_Sonde_Name}: The sonde name for the sonde in the water at the end of the visit.
\item \strong{Removal_Time}: The time the sonde was removed from the water.
\item \strong{Return_Time}: The time the sonde was returned to the water.
\item \strong{Next_Timepoint}: The next timepoint that data will be collected.
This should be the next "good" time point after the sonde has been returned to the water, but is sometimes the timepoint
when the wiper is checked and the sonde is out of the water.
\item \strong{Data_Download}: Logical, was data downloaded at this visit?
\item \strong{Download_Device}: The name of the device data was downloaded to.
\item \strong{Remove_Period}: Logical, is there a data disruption that merits removing data during the out of water period. Used to skip periods
where we have coarse out of water periods (missing times), but the data appears uninterupted to prevent removing excess data.
\item \strong{Crew}: The names or initials of the people performing the field visit.
\item \strong{Weather}: A description of the weather during the field visit.
\item \strong{Notes}: Notes associated with the field visit.
\item \strong{DataEntry_Notes}: Additional notes from data entry. Can have any name that includes "Notes"
}
}
\description{
Loads a \code{.csv} file with field visit metadata. Ensures correct column classes and cleans up
extra notes.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "example-fieldform.csv")
fieldform <- read_ff(file)
}
