% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sonde.R
\name{read_sonde}
\alias{read_sonde}
\title{Upload raw sonde data}
\usage{
read_sonde(
  file,
  return = "df",
  encoding = NULL,
  flags = FALSE,
  skip = NULL,
  tz = "Etc/GMT+8"
)
}
\arguments{
\item{file}{File path to sonde data (.csv).}

\item{return}{What format should data be returned in? Either \code{df} or \code{sonde}.}

\item{encoding}{File encoding, will guess if \code{NULL}.}

\item{flags}{Logical, if \code{TRUE} a flag column will be added for each parameter.}

\item{skip}{The number of rows to skip, assumes the first row is a header,  will guess if \code{NULL}.}

\item{tz}{Time zone for the file. If not specified will use user timezone. See details for more information.}
}
\value{
If \code{return} is \code{df}:
\strong{A \code{data.frame} containing the date, time, site name, and the "core" measurements (when available):}
\itemize{
\item Specific conductivity measured in µS/cm.
\item Fluorescent dissolved organic matter (fDOM) measured in Quinine Sulfate Units (QSU).
\item Dissolved oxygen measured in mg/L.
\item Turbidity measured in Formazin Nephelometric Units (FNU).
\item pH measured in pH units.
\item Temperature measured in degrees C.
\item Battery voltage measured in volts.
}

If \code{return} is \code{sonde}:
\strong{An object of class \code{sonde} containing:}
\itemize{
\item \strong{file}: the file name of the sonde data.
\item \strong{serials}: the probe serial numbers associated with each measurement.
\item \strong{data}: the sonde data formatted as described above.
}
}
\description{
Loads a \code{.csv} file with raw sonde data. Turns into a \code{sonde} object.
Cleans column names, adds a \code{DateTime} column, and optionally adds a \code{flag} column.
}
\details{
Daylights saving time can cause issues for continuous datasets, so data is often collected in standard time.
If this is the case use Etc/GMT offsets. See \link[base]{OlsonNames} for all available timezones. Note that
Etc/GMT signs are \strong{reverse} of UTC signs. For example UTC-8 would be Etc/GMT+8,
signifying Pacific Standard Time (PST).
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "example-csv-data1.csv")
data <- read_sonde(file)

file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-usb-example.csv")
data <- read_sonde(file, return = "sonde")
}
