% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_flagged}
\alias{remove_flagged}
\title{Set flagged values to missing}
\usage{
remove_flagged(df, flag_names)
}
\arguments{
\item{df}{the `data.frame` to remove flagged values from}

\item{flag_names}{the flag names to set to `NA` if the flag is `TRUE`}
}
\value{
a `data.frame` with the flagged values removed
}
\description{
Set flagged values to missing
}
