% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Run the SondePolishR App}
\usage{
run_app(default_path = getwd())
}
\arguments{
\item{default_path}{Default filepath used for saving data to, defaults to current working directory.}
}
\value{
Shiny App
}
\description{
Opens a window with the interactive Shiny App to
interactively load, view, correct, and export sonde data.
}
\examples{
\dontrun{
if(interactive()){
  library(SondePolishR)
  run_app()
}}
}
