% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-interp.R
\name{run_interp}
\alias{run_interp}
\title{Interpolate data gaps}
\usage{
run_interp(data_interp, y_var, method, freq = 1)
}
\arguments{
\item{data_interp}{\code{data.frame} from \code{prep_interp()} with dupes condensed.}

\item{y_var}{Variable being interpolated.}

\item{method}{The method to use for interpolation, options include:
\code{linear}, \code{spline}, \code{random_forest}, and \code{ts_interp} see details for further information about the methods.}

\item{freq}{The period interval for the time series in days, only used if \code{method} is \code{ts_interp}.}
}
\value{
a \code{data_interp} with a extra column added:
\itemize{
\item \code{yvar_fill}: Filled values for \code{y_var}.
}
}
\description{
Uses the specified method to fill NA values for a given parameter within the dataset.
}
\details{
Currently the follow interpolation options are supported:
\itemize{
\item \code{linear}: Uses the zoo::na.approx() function to linearly fill gaps.
\item \code{spline}: Uses the zoo::na.spline() function to fill gaps via spline interpolation (makes smoother curves).
\item \code{random_forest}: Depending on how much data is in the project, this option can be a little slow. Uses missForest::missForest() to
fill gaps. uses responses across all the available parameters to try and fill in the missing parameter.
\item \code{ts_interp}: Uses forecast::na.interp() which treats the data as a time series and make predictions based on the natural seasonal
patterns within the data. The seasonality is determined by \code{freq}. For instance, if set to 1, it will account for daily fluctuations.
If set to 365 it will look at annual fluctuations.
}
}
\examples{
interp_df <- prep_interp(example_sondeproj)
filled_yvar <- run_interp(interp_df, "fDOM_QSU", "linear")
}
