% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift-points.R
\name{shift_points}
\alias{shift_points}
\title{Correct points via an absolute shift}
\usage{
shift_points(data, par, index, shift_val = NULL)
}
\arguments{
\item{data}{a data.frame with sonde data}

\item{par}{the parameter being corrected}

\item{index}{the index values of the rows that need to be shifted}

\item{shift_val}{a list of the slope and int (intercept) to use to shift the data by, if \code{NULL}, it will be guessed using \link[SondePolishR]{guess_shift}}
}
\value{
a \code{data.frame} with the values adjusted
}
\description{
Sometimes a continuous sonde signal will be shifted up or down by a consistent value for a period of time, this will adjust the
dataset for those values by that set amount.
}
\examples{
example_sondeproj$data$ODO_mg_L[5:7]
data <- shift_points(example_sondeproj$data, "ODO_mg_L", 5:7)
data$ODO_mg_L[5:7]
}
