% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SondePolishR-package.R
\docType{package}
\name{SondePolishR-package}
\alias{SondePolishR}
\alias{SondePolishR-package}
\title{SondePolishR: Perform Quality Assurance and Data Corrections on Sonde Data}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Katie Wampler \email{katie.wampler@oregonstate.edu} (\href{https://orcid.org/0000-0002-5584-7554}{ORCID})

Authors:
\itemize{
  \item Katie Wampler \email{katie.wampler@oregonstate.edu} (\href{https://orcid.org/0000-0002-5584-7554}{ORCID})
}

}
\keyword{internal}
