% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-func.R
\name{update_limits}
\alias{update_limits}
\title{Update numeric input with limits from sonde data}
\usage{
update_limits(df, y_var, session)
}
\arguments{
\item{df}{data.frame with sonde data}

\item{y_var}{parameter to plot}

\item{session}{the environment that can be used to access information and functionality relating to the session}
}
\description{
Update numeric input with limits from sonde data
}
