% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usb_export}
\alias{usb_export}
\title{Detect if file was downloaded via a USB or not}
\usage{
usb_export(file)
}
\arguments{
\item{file}{file path to sonde data (.xlsx or .csv)}
}
\value{
Logical, if TRUE file has USB download structure, if FALSE file does not.
}
\description{
Sonde data files differ based on how the data is downloaded, this is used to detect so
data can be processed correctly.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-example.csv")
usb_export(file)

file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-usb-example.csv")
usb_export(file)
}
