% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{commit_diff}
\alias{commit_diff}
\alias{apply_diff}
\title{Get and apply changes to data}
\usage{
commit_diff(olddata, newdata)

apply_diff(olddata, diff)
}
\arguments{
\item{olddata}{The original \code{data.frame} to compare to.}

\item{newdata}{The new \code{data.frame} you want to get the changes for.}

\item{diff}{A \code{data_diff} object storing the difference between \code{olddata} and \code{newdata} or a list of \code{data_diff}.}
}
\value{
\code{commit_diff} returns a \code{data_diff} object storing the differences between \code{data.frames}. See \code{daff} package for more details.
\code{apply_diff} returns a \code{data.frame} with the \code{diff} applied.
}
\description{
Wrappers for \code{daff::diff_data()} and \code{daff::apply_patch()}. Natively \code{daff} functions don't play
nicely with dates and datetimes. So the wrapper confirms they're the same between data and
then temporarily removes them to get the diff without those columns. Used to succinctly track changes to
the raw data and revert changes when needed.
}
\examples{
data_edit <- example_data

#make a change in newdata
data_edit$fDOM_QSU[1:4] <- NA
head(data_edit)

#save difference
dd <- commit_diff(example_data, data_edit)

#apply difference
data_edit2 <- apply_diff(example_data, dd)
head(data_edit2)
}
