% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_1-weekly-dates-module.R
\name{weekly_range_sidebar_UI}
\alias{weekly_range_sidebar_UI}
\alias{weekly_range_buttons_UI}
\alias{weekly_range_server}
\title{View data by a specified time length}
\usage{
weekly_range_sidebar_UI(id)

weekly_range_buttons_UI(id)

weekly_range_server(id, sondeproj, data_ver, view_state)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{view_state}{A `reactiveVal` holding a list of items specifying the view state:
- dates: The range of dates being viewed via the date selector
- period_view: Logical if the period view is being used
- period_length: Length of period view
- period_n: The period number to view.}
}
\value{
a reactive of length two with the min and max dates.
}
\description{
Used to adjust plotting to show periods to better examine details.
}
\keyword{internal}
