% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-flag.R
\name{add_flags}
\alias{add_flags}
\title{Add new flag to dataset}
\usage{
add_flags(data, y_var, index, flag)
}
\arguments{
\item{data}{Data.frame with sonde data.}

\item{y_var}{The column to add flag to.}

\item{index}{Row numbers to add flag to.}

\item{flag}{Flag to add to the data.frame.}
}
\value{
a \code{data.frame} with the same dimensions as \code{data} with the flags added to the appropriate column.
}
\description{
Safely adds a flag to a specific parameter without overwriting old flags. Also maintains a consistent order
so that additional flags can be identified with version control.
}
\examples{
data <- add_flags(example_sondeproj$data, "fDOM_QSU", 2:7, "TEST01")
data$fDOM_QSU_flag[1:8]
}
