% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{add_flags}
\alias{add_flags}
\title{Get skeleton flagging dataframe}
\usage{
add_flags(proj, data)
}
\arguments{
\item{proj}{a \code{sondeproj} object to add the flags to. If it has existing flags they will be merged
with the new flags from \code{data} with existing flags.}

\item{data}{a data.frame with sonde data}
}
\value{
a data.frame
\itemize{
\item if \code{par} and \code{flag_name} are \code{NULL} it will return a \code{data.frame} with the same number of rows as \code{data} but
with a blank column for each parameter in the \code{data.frame} only the index, datetimme, datetime_rd, and DupNum columns.
}
}
\description{
Works to provide a skeleton for the flag data (if no parameter or flag names are provided).
}
\examples{
#add flag columns
updated_proj <- add_flags(example_sondeproj, example_data)
colnames(updated_proj$flags$flag_rm)
}
