% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{add_flags}
\alias{add_flags}
\title{Add flags to sonde data}
\usage{
add_flags(data, par = NULL, flag_name = NULL, index = NULL, makeNA = TRUE)
}
\arguments{
\item{data}{a data.frame with sonde data}

\item{par}{the parameter to check}

\item{flag_name}{a character with the name of the flag}

\item{index}{the index values for the rows to be flagged}

\item{makeNA}{if \code{TRUE} the flagged data will be converted to NA values}
}
\value{
a data.frame
\itemize{
\item if \code{par} and \code{flag_name} are \code{NULL} it will return a new column for each parameter in the data.frame with the form <*_flag>
\item if \code{par} and \code{flag_name} are specified it will add a named vector to the <par_flag> column specifying \code{TRUE} for flagged and \code{FALSE} for not flagged
}
}
\description{
Works either to provide a skeleton for the flag data (if no parameter or flag names are provided) or
adds a flag to the specified data. Used to keep track of which datapoints were flagged in which QA/QC steps.
If only some flags are added, it will try to add the missing flags.
}
\examples{
#add flag columns
data <- add_flags(raw_sonde)
colnames(data)

#add a flag
data <- add_flags(raw_sonde, "fDOM_QSU", "test_flag", c(1,2,3))
head(data)
}
