% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8_1-correction-methods.R
\name{additive_UI}
\alias{additive_UI}
\alias{additive_server}
\title{Apply an additive shift to data}
\usage{
additive_UI(id)

additive_server(
  id,
  sondeproj,
  y_var,
  plot,
  plot_data,
  currplot,
  curredit,
  currmethod,
  index
)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{y_var}{A `reactiveVal` holding the y-variable being plotted.}

\item{plot}{A `reactiveVal` holding a basic plot before things were added.}

\item{plot_data}{A `reactiveVal` holding the current data being plotted.}

\item{currplot}{A `reactiveVal` holding the current main plot.}

\item{curredit}{A `reactiveVal` holding the current edit object.}

\item{currmethod}{A `reactiveVal` holding the current correction method.}

\item{index}{A `reactiveVal` holding the indices of the selected points within the full dataset.}
}
\value{
Updates the values of `currplot` and `curredit` to be current with the method.
}
\description{
Used to adjust data either by a single value or apply a linear correction.
}
\keyword{internal}
