% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8-additive.R
\name{additive_UI}
\alias{additive_UI}
\alias{additive_server}
\title{Address any additive shifts}
\usage{
additive_UI(id)

additive_server(id, sondeproj, data_ver, y_var)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset, user can select a group of points and apply a additive shift to the data to correct for shifts.
}
\keyword{internal}
