% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_1-logging-data-change.R
\name{apply_edit_UI}
\alias{apply_edit_UI}
\alias{apply_edit_server}
\title{Shiny module to added flags and save changes to  data}
\usage{
apply_edit_UI(id, note = NULL)

apply_edit_server(id, sondeproj, edit)
}
\arguments{
\item{id}{the shiny ID of the action button}

\item{sondeproj}{A `reactiveVal` holding the current `sondeproj`.}

\item{edit}{A `reactiveVal` holding a list of length six:
- data: new updated data as a `data.frame`
- rows: logical vector which specifies rows changed a TRUE
- y_var: parameter being edited
- step: name of the editing step for the changelog
- note: an optional note to add to the changelog
- flag: character flag to use for edits to the data}
}
\description{
The UI creates a button to allow user to save the changes made to the data. The server function will provide a message to the user to
let them know the changes (or lack of changes made), add flags to the dataset, and save the dataset as a new version.
}
\keyword{internal}
