% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{apply_diff}
\alias{apply_diff}
\title{Apply a diff object to a dataset}
\usage{
apply_diff(
  data,
  diff,
  id = c("DateTime_rd", "DupNum"),
  invert = FALSE,
  skip_merge = TRUE
)
}
\arguments{
\item{data}{the data to apply the \code{diff} to. Must contain all the columns in \code{diff} or a list of flags.}

\item{diff}{a list of \code{diff} or a single \code{diff} objects generated using \code{get_diff}.}

\item{id}{name of the column name used to match observations between \code{olddata} and \code{newdata}.}

\item{invert}{logical. If \code{TRUE} changes will be reversed.}

\item{skip_merge}{logical. If \code{TRUE} will skip any \code{diff} with are data merges.}
}
\value{
a \code{data.frame} with the same columns as \code{data} with the changes from \code{diff} applied.
Note that this could increase the number of rows if diff is a data merge.
}
\description{
Once changes between \code{data.frame}s have been saved as a \code{diff} object, they can be used to move between
the changes made by applying the \code{diff} to data.
}
\examples{
data1 <- example_data[1:10,]
data2 <- data1
data2$fDOM_QSU[1:4] <- NA
diff <- get_diff(data1, data2)

#get new data from original
newdata2 <- apply_diff(data1, diff)
all.equal(data2, newdata2)

#get orginal data from new data
newdata1 <- apply_diff(data2, diff, invert=TRUE)
all.equal(data1, newdata1)

}
