% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{apply_diff}
\alias{apply_diff}
\title{Apply a diff object to a dataset}
\usage{
apply_diff(data, diff, id = "DateTime", invert = FALSE, skip_merge = TRUE)
}
\arguments{
\item{data}{the data to apply the `diff` to. Must contain all the columns in `diff`.}

\item{diff}{a list of `diff` or a single `diff` objects generated using `get_diff`.}

\item{id}{name of the column name used to match observations between `olddata` and `newdata`.}

\item{invert}{logical. If `TRUE` changes will be reversed.}

\item{skip_merge}{logical. If `TRUE` will skip any `diff` with are data merges.}
}
\value{
a `data.frame` with the same columns as `data` with the changes from `diff` applied.
Note that this could increase the number of rows if diff is a data merge.
}
\description{
Once changes between `data.frame`s have been saved as a `diff` object, they can be used to move between
the changes made by applying the `diff` to data.
}
\examples{
data1 <- example_data[1:10,]
data2 <- data1
data2$fDOM_QSU[1:4] <- NA
diff <- get_diff(data1, data2)

#get new data from original
newdata2 <- apply_diff(data1, diff)
all.equal(data2, newdata2)

#get orginal data from new data
newdata1 <- apply_diff(data2, diff, invert=TRUE)
all.equal(data1, newdata1)

}
