% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-edit.R
\name{apply_edit}
\alias{apply_edit}
\title{Log edits to a \code{sondeproj}}
\usage{
apply_edit(proj, edit, username)
}
\arguments{
\item{proj}{A \code{sondeproj} object holding sonde data.}

\item{edit}{A list of length six:
\itemize{
\item data: new updated data as a \code{data.frame}
\item rows: vector of indices that were changed
\item y_var: parameter being edited
\item step: name of the editing step for the changelog
\item note: an optional note to add to the changelog
\item flag: character flag to use for edits to the data
}}

\item{username}{the username of the person who made the change}
}
\value{
A \code{sondeproj} object with edits made.
}
\description{
Uses a list of edit parameters to update a \code{sondeproj} with the specified edits. Makes
changes to the data, changelog, and flags.
}
\examples{
data <- example_data
data$fDOM_QSU[1:4] <- NA
rows <- c(1:4)
rows[1:4] <- TRUE
edit <- list(data = example_data,
             rows = rows,
             y_var = "fDOM_QSU",
             step = "outlier removal",
             note = "example edit",
             flag = "RM07")
updated_proj <- apply_edit(example_sondeproj, edit, "Smith")

}
