% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj-tracking.R
\name{get_log}
\alias{get_log}
\alias{write_log}
\alias{clear_log}
\alias{set_log}
\title{Get and Write to Change Log}
\usage{
get_log(env = .pkgenv)

write_log(
  par,
  step,
  n,
  note = "",
  version,
  datetime = Sys.time(),
  user = Sys.info()[["user"]],
  env = .pkgenv
)

clear_log(env = .pkgenv)

set_log(log, env = .pkgenv)
}
\arguments{
\item{env}{the environment in which configuration settings is stored. Defaults to the package environment.}

\item{par}{the name of the parameter modified}

\item{step}{a description of the type of change made}

\item{n}{the number of points modified}

\item{note}{a note from the analyst about the change made}

\item{version}{the associated file version name}

\item{datetime}{the date and time the change was made}

\item{user}{the username of the person who made the change}

\item{log}{a existing change log read in from read_project}
}
\value{
a \code{data.frame} with the log from the package environment
}
\description{
Log is a \code{data.frame} stored in the package environment used to keep track of the changes
made to the initial raw sonde data.
}
\examples{
get_log()

write_log("Cond_S_cm", "physical limits", 5, "making an example", "V1")
}
