% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj-tracking.R
\name{get_log}
\alias{get_log}
\alias{write_log}
\title{Get and Write to Change Log}
\usage{
get_log(sondeproj)

write_log(
  sondeproj,
  par,
  step,
  n,
  note = "",
  diff_name = NULL,
  datetime = Sys.time(),
  user = Sys.info()[["user"]],
  return = "df"
)
}
\arguments{
\item{sondeproj}{the sonde project to get the change log from. If \code{NULL} will return a blank change log structure.}

\item{par}{the name of the parameter modified}

\item{step}{a description of the type of change made}

\item{n}{the number of points modified}

\item{note}{a note from the analyst about the change made}

\item{diff_name}{the associated diff object}

\item{datetime}{the date and time the change was made}

\item{user}{the username of the person who made the change}

\item{return}{either \code{df} or \code{sondeproj} to specify if the changelog only should be returned or the \code{sondeproj} with
the change log updated.}
}
\value{
a \code{data.frame} with the log from \code{sondeproj}
}
\description{
The change log is a component of a \code{sondeproj} object used to keep track of the changes
made to the initial raw sonde data.
}
\examples{
write_log(NULL, "Cond_S_cm", "physical limits", 5, "making an example", "diff1")
}
