% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3-check-data.R
\name{check_data_UI}
\alias{check_data_UI}
\alias{check_data_server}
\title{Check data for duplicates, gaps, remove OOW periods}
\usage{
check_data_UI(id)

check_data_server(id, sondeproj, data_ver, y_var)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}
}
\value{
Invisible NULL
}
\description{
Check data for duplicates, gaps, remove OOW periods
}
\keyword{internal}
