% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8-corrections.R
\name{correction_UI}
\alias{correction_UI}
\alias{correction_server}
\title{Address any data shifts or corrections}
\usage{
correction_UI(id)

correction_server(id, sondeproj, data_ver, y_var, view_state, username)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{view_state}{A \code{reactiveVal} holding a list of items specifying the view state:
\itemize{
\item abs_dates: The absolute range of dates within the dataset
\item dates: The range of dates being viewed via the date selector
\item period_view: Logical if the period view is being used
\item period_length: Length of period view
\item period_n: The period number to view.
}}

\item{username}{A \code{reactiveVal} holding the name of the analyst for the changelog}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset, user can select a group of points and apply a additive shift to the data to correct for shifts, apply
a drift correction to a data file, or apply a smoothing function to selected data.
}
\keyword{internal}
