% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8-corrections.R
\name{correction_UI}
\alias{correction_UI}
\alias{correction_server}
\title{Address any data shifts or corrections}
\usage{
correction_UI(id)

correction_server(id, sondeproj, data_ver, y_var, period_view, dates, p_length)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{period_view}{Should data be viewed by period?}

\item{dates}{The date range to view the data.}

\item{p_length}{The length of the period to view.}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset, user can select a group of points and apply a additive shift to the data to correct for shifts, apply
a drift correction to a data file, or apply a smoothing function to selected data.
}
\keyword{internal}
