% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step8_1-correction-methods.R
\name{drift_UI}
\alias{drift_UI}
\alias{drift_server}
\title{Apply a drift correction to a data file}
\usage{
drift_UI(id, sondeproj)

drift_server(
  id,
  sondeproj,
  y_var,
  plot,
  plot_data,
  currplot,
  curredit,
  currmethod
)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{y_var}{A `reactiveVal` holding the y-variable being plotted.}

\item{plot}{A `reactiveVal` holding a basic plot before things were added.}

\item{plot_data}{A `reactiveVal` holding the current data being plotted.}

\item{currplot}{A `reactiveVal` holding the current main plot.}

\item{curredit}{A `reactiveVal` holding the current edit object.}

\item{currmethod}{A `reactiveVal` holding the current correction method.}
}
\value{
Updates the values of `currplot` and `curredit` to be current with the method.
}
\description{
Used to account for instrument drift by applying a linear correction to a data file. Uses differences between
check and resident sonde measurments when available as the default correction amount.
}
\keyword{internal}
