% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Sample Sonde Data}
\format{
An data.frame object with 8071 rows and 13 columns:
\itemize{
\item \strong{Index}: A number, 1 through the number of rows in the \code{data.frame}.
\item \strong{Date}: Date of measurement in the form YYYY-MM-DD.
\item \strong{Time_HH_mm_ss}: Time of measurement in the form hh:mm:ss.
\item \strong{DateTime}: Date and time of measurement as a POSIXct.
\item \strong{DateTime_rd}: Date and time of measurement as a POSIXct rounded to the nearest data interval value.
\item \strong{Site_Name}: Name of the site.
\item \strong{Battery_V}: Battery voltage when measurement was collected.
\item \strong{fDOM_QSU}: Fluorescent dissolved organic matter (fDOM) measured in Quinine Sulfate Units (QSU).
\item \strong{ODO_mg_L}: Dissolved oxygen measured in mg/L.
\item \strong{pH}: Dissolved oxygen measured in pH units.
\item \strong{SpCond_uS_cm}: Specific conductivity measured in µS/cm.
\item \strong{Temp_C}: Temperature measured in degrees C.
\item \strong{Turbidity_FNU}: Turbidity measured in Formazin Nephelometric Units (FNU).
}
}
\source{
Forest Ecohydrology and Watershed Science Lab (2024) Fall Creek: 2024-07-31 to 2024-10-23. Data set.
}
\usage{
example_data
}
\description{
An example of the output from \link[SondePolishR]{read_sonde} function.
}
\keyword{datasets}
