% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2-explore-plot.R
\name{explore_data_UI}
\alias{explore_data_UI}
\alias{explore_data_server}
\title{Visualize dataset by analyte}
\usage{
explore_data_UI(id)

explore_data_server(id, sondeproj, data_ver, y_var)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}
}
\value{
Invisible NULL
}
\description{
Plots loaded dataset with options to explore full dataset or view the data in weekly sections.
Plots are created using \code{plotly} so they are interactive. Module also allows the user to view
the changes over the dataset versions via row selection in a table via the \code{log}.
}
\keyword{internal}
