% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step10-export-data.R
\name{export_UI}
\alias{export_UI}
\alias{export_server}
\title{Export data and metadata}
\usage{
export_UI(id)

export_server(id, sondeproj, data_ver, y_var, current_mod, webgl_supported)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{current_mod}{The name of the current module being viewed.}

\item{webgl_supported}{A `reactiveVal` indicating if webgl is support in the current browser.}
}
\description{
Save to file the corrected data and metadata including summaries of the data.
}
\keyword{internal}
