% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step9-fDOM.R
\name{fdom_UI}
\alias{fdom_UI}
\alias{fdom_server}
\title{Correct fDOM data for temperature and turbidity}
\usage{
fdom_UI(id)

fdom_server(id, sondeproj, data_ver, y_var, view_state, username)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{view_state}{A `reactiveVal` holding a list of items specifying the view state:
- abs_dates: The absolute range of dates within the dataset
- dates: The range of dates being viewed via the date selector
- period_view: Logical if the period view is being used
- period_length: Length of period view
- period_n: The period number to view.}

\item{username}{A `reactiveVal` holding the name of the analyst for the changelog}
}
\description{
The fDOM signal can be affected but sediment and temperature, but these effects can be corrected for.
}
\keyword{internal}
