% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{flag_data}
\alias{flag_data}
\title{Flag dataset and save a copy}
\usage{
flag_data(df, par, flag_name, index, prj_path)
}
\arguments{
\item{df}{a data.frame with sonde data}

\item{par}{the parameter to check}

\item{flag_name}{a character with the name of the flag}

\item{index}{the index values for the rows to be flagged}

\item{prj_path}{the file path to save the sonde project to}
}
\value{
returns the flagged copy of the \code{data.frame}
}
\description{
Adds flags to the dataset, saves the new version of the dataset, and saves the project to the
project path specified by the user as a \code{.qs} object. Also provides messages about the
changes being made.
}
\examples{
prj_path <- file.path(withr::local_tempdir(), "test_prj.qs")

#do intial versioning things (automatic in app)
  clear_data()
  clear_log()
  write_data(raw_sonde, "raw")

#flag data
  df <- flag_data(raw_sonde, "fDOM_QSU", "test_flag", 1:4, prj_path)
}
