% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-eco-limits.R
\name{get_eco_limits}
\alias{get_eco_limits}
\title{Get USGS Sonde Parameter Limits}
\usage{
get_eco_limits(ecoregion, parameter)
}
\arguments{
\item{ecoregion}{Name for Level III North America Ecoregion (NA_L3NAME)}

\item{parameter}{USGS parameter code(s) for the parameters you'd like to get limits of}
}
\value{
A data.frame for each parameter code that has two rows and 10 columns:
\itemize{
\item statistic_id: 01 is max daily value, 02 is min daily value
\item max: the maximum of the daily min or max values
\item q999: the 99.9th quantile of the daily min or max values
\item q99: the 99th quantile of the daily min or max values
\item q95: the 95th quantile of the daily min or max values
\item q05: the 5th quantile of the daily min or max values
\item q01: the 1st quantile of the daily min or max values
\item q001: the 0.1th quantile of the daily min or max values
\item min: the minimum of the daily min or max values
\item n: the number of daily values used to calculate metrics
\item par: the USGS parameter code
\item stat_n: the number of USGS stations used to calculate metrics
}

If multiple parameters are requested, it will return a list of data.frames
}
\description{
Pulls daily water quality records from USGS for stations within the specified ecoregion and returns
the distribution of the minimum and maximum values for use in sonde QA-QC.
}
\details{
Potential parameter code common for USGS continuous water quality:
\itemize{
\item 00010: Temperature, water, degrees Celsius
\item 00095: Specific conductance, water, unfiltered, microsiemens per centimeter at 25 degrees Celsius
\item 00300: Dissolved oxygen, water, unfiltered, milligrams per liter
\item 00301: Dissolved oxygen, water, unfiltered, percent of saturation
\item 00400: pH, water, unfiltered, field, standard units
\item 00480: Salinity, water, unfiltered, parts per thousand
\item 32295: Dissolved organic matter fluorescence (fDOM), water, in situ, concentration estimated from reference material, micrograms per liter as quinine sulfate equivalents (QSE)
\item 32322: Dissolved organic matter relative fluorescence (fDOM), water, in situ, relative fluorescence units (RFU)
\item 63680: Turbidity, water, unfiltered, monochrome near infra-red LED light, 780-900 nm, detection angle 90 +-2.5 degrees, formazin nephelometric units (FNU)
\item 72147: Depth of sensor below water surface, feet
\item 99409: Suspended sediment concentration, water, unfiltered, estimated by regression equation, milligrams per liter
}
}
\note{
You are likely to run into API limits with this function unless you request an \href{https://api.waterdata.usgs.gov/signup/}{API} from USGS

If a ecoregion returns more than 300 sites for a particular parameter, it will randomly sample 500 of the sites to
determine the limits, to conserve computing power and USGS resources.
}
\examples{
get_eco_limits("Blue Mountains", "32295")
}
