% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-precip.R
\name{get_precip}
\alias{get_precip}
\title{Get hourly precipitation at data site}
\usage{
get_precip(data, lat, long)
}
\arguments{
\item{data}{the data to get matching precipitation data for}

\item{lat}{the latitude to get precipitation data at}

\item{long}{the longitude to get precipitation data at}
}
\value{
a data.frame with two columns:
\itemize{
\item DateTime: The datetime (\code{POSIXct}) in the same timezone as the data, at an hourly resolution.
\item Precip_mm_hr: Average MERRA-2 bias corrected total precipitation at the surface of the earth in mm per hour.
}
}
\description{
Precipitation is downloaded from \href{https://power.larc.nasa.gov/}{NASA Power} at an hourly scale based on the provided
coordinates.
}
\examples{
data <- example_data \%>\% filter(Date == "2024-08-01")
get_precip(data, 43.96775, -122.63012)
}
