% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get raw data from sonde project}
\usage{
get_raw_data(proj)
}
\arguments{
\item{proj}{A `sondeproj` object.}
}
\value{
a `data.frame` with raw data.
}
\description{
Includes any raw data merges and returns the raw data without any changes. Useful for viewing data before changes were made.
Will include all raw data, even if additional datasets were merged partway through cleaning.
}
\examples{
raw <- get_raw_data(example_sondeproj)
}
