% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift-points.R
\name{guess_shift}
\alias{guess_shift}
\title{Guess the amount to shift observations}
\usage{
guess_shift(data, par, index)
}
\arguments{
\item{data}{a data.frame with sonde data}

\item{par}{the parameter being corrected}

\item{index}{the index values of the rows that need to be shifted}
}
\value{
a numeric with the guessed shift value based on the parameter
}
\description{
Sometimes a continuous sonde signal will be shifted up or down by a consistent value for a period of time. This function will use the
data before and after the selected observations to guess the appropriate value to correct the data by
}
\examples{
guess_shift(example_sondeproj$data, "ODO_mg_L", 5:7)
}
