% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-gaps.R
\name{identify_gaps}
\alias{identify_gaps}
\title{Identify missing observations with sonde data}
\usage{
identify_gaps(data, ignore = NULL)
}
\arguments{
\item{data}{a \code{data.frame} with sonde data.}

\item{ignore}{the length in minutes to ignore gaps}
}
\value{
a \code{data.frame} with the following columns with a row for each missing period:
\itemize{
\item start: starting datetime of the missing data
\item end: ending datetime of the missing data
\item gap_length: number of points in the missing section
\item user_note: user input information about the duplicated section
}
}
\description{
Uses the interval-rounded datetime (\code{DateTime_rd}) observations to identify periods of missing
observations.
}
\examples{
identify_gaps(example_data, ignore=0)
}
