% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7-interp.R
\name{interp_UI}
\alias{interp_UI}
\alias{interp_server}
\title{Fill in data gaps using interpolation}
\usage{
interp_UI(id)

interp_server(
  id,
  sondeproj,
  data_ver,
  y_var,
  period_view,
  dates,
  p_length,
  current_mod
)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{period_view}{Should data be viewed by period?}

\item{dates}{The date range to view the data.}

\item{p_length}{The length of the period to view.}

\item{current_mod}{The name of the current module being viewed.}
}
\description{
Attempts to fill gaps less than the max length using the specified interpolation method.
}
\keyword{internal}
