% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step5-limits.R
\name{limits_UI}
\alias{limits_UI}
\alias{limits_server}
\title{Flagging data the is outside specified limits}
\usage{
limits_UI(id)

limits_server(id, sondeproj, data_ver, y_var, period_view, dates, p_length)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{period_view}{Should data be viewed by period?}

\item{dates}{The date range to view the data.}

\item{p_length}{The length of the period to view.}
}
\description{
There are certain thresholds for some of the sonde parameters that aren't physical possible (i.e, water temperature above 100 deg C).
This module visualizes those limits and flags data outside specified limits.
}
\keyword{internal}
