% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step3-limits.R
\name{limits_UI}
\alias{limits_UI}
\alias{limits_server}
\title{Flagging data the is outside specified limits}
\usage{
limits_UI(id)

limits_server(id, data)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{data}{A reactive holding the loaded dataset.}
}
\description{
There are certain thresholds for some of the sonde parameters that aren't physical possible (i.e, water temperature above 100 deg C).
This module visualizes those limits and flags data outside specified limits.
}
\keyword{internal}
