% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1-load-data.R
\name{load_data_UI}
\alias{load_data_UI}
\alias{load_data_server}
\title{Read in the dataset or project and set save path}
\usage{
load_data_UI(id)

load_data_server(id, sondeproj, data_ver, view_state, username)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{data_ver}{A \code{reactiveVal} holding a number used to track when new data is added to trigger resets.}

\item{view_state}{A \code{reactiveVal} holding a list of items specifying the view state:
\itemize{
\item abs_dates: The absolute range of dates within the dataset
\item dates: The range of dates being viewed via the date selector
\item period_view: Logical if the period view is being used
\item period_length: Length of period view
\item period_n: The period number to view.
}}

\item{username}{A \code{reactiveVal} holding the name of the analyst for the changelog}
}
\value{
The loaded data as a reactive object.
}
\description{
Takes in a dataset as a \code{.csv} or a sonde project as an \code{.RDS} file via file selection.
If the data is a sonde project the save path will default it it's existing path, otherwise
the user will need to select a save path with the file name based on the name of the data file.
}
\keyword{internal}
