% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1-load-data.R
\name{load_data_UI}
\alias{load_data_UI}
\alias{load_data_server}
\title{Read in the dataset or project and set save path}
\usage{
load_data_UI(id)

load_data_server(id)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}
}
\value{
The loaded data as a reactive object.
}
\description{
Takes in a dataset as a \code{.csv} or a sonde project as an \code{.RDS} file via file selection.
If the data is a sonde project the save path will default it it's existing path, otherwise
the user will need to select a save path with the file name based on the name of the data file.
}
\keyword{internal}
