% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_4-main-plot-module.R
\name{main_plot_UI}
\alias{main_plot_UI}
\alias{main_plot_server}
\title{Visualize main data plot}
\usage{
main_plot_UI(id)

main_plot_server(
  id,
  sondeproj,
  plot_obj,
  plot_data,
  y_var,
  plot_name = NULL,
  sel_mode = FALSE,
  plot_exist = reactiveVal(),
  startmin = reactiveVal(),
  startmax = reactiveVal()
)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{plot_obj}{Plotly object to plot.}

\item{plot_data}{A `reactiveVal` holding the current dataset to plot.}

\item{y_var}{A `reactiveVal` holding the Y-variable to plot on the y-axis.}

\item{plot_name}{The plot name to register the plot as for selecting points}

\item{sel_mode}{Logical, should selection mode be turned on as default?}

\item{plot_exist}{A `reactiveVal` indicating if the plot exists or not to prevent warnings about plot obj not being registered.}
}
\value{
a plot of the data.
}
\description{
Comes with an adjustable y-axis to adjust the values being viewed.
}
\keyword{internal}
