% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version-control.R
\name{new_version}
\alias{new_version}
\title{Checks if the data is different than the current saved version}
\usage{
new_version(df)
}
\arguments{
\item{df}{the data.frame to check to see if it's new}
}
\value{
TRUE if df differs from the previous version,
FALSE if df is the same as the previous version
}
\description{
Checks if the data is different than the current saved version
}
