% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step6-outliers.R
\name{outlier_UI}
\alias{outlier_UI}
\alias{outlier_server}
\title{Flag data identified as outliers either manually or view methods}
\usage{
outlier_UI(id)

outlier_server(id, sondeproj, data_ver, y_var, period_view, dates, p_length)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{period_view}{Should data be viewed by period?}

\item{dates}{The date range to view the data.}

\item{p_length}{The length of the period to view.}
}
\description{
Looks for "weird" data where there are large spikes within a short period that are likely unrealistic and caused by
instrument malfunction or a bubble near the sensor.
}
\keyword{internal}
