% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2-plotting-opts.R
\name{plot_options_UI}
\alias{plot_options_UI}
\alias{plot_options_server}
\title{Specify Plotting Options}
\usage{
plot_options_UI(id)

plot_options_server(id)
}
\arguments{
\item{id}{the shiny ID of the module}
}
\value{
a list of length 5:
\itemize{
\item points: should points be plotted?
\item line: should line be plotted?
-files: should points be colored by file?
-oow: should out of water periods be plotted?
-calcheck: should cal check be plotted?
}
}
\description{
Used to create and return selections about what to include in the plot.
}
\keyword{internal}
