% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sonde-data.R
\name{plot_sonde}
\alias{plot_sonde}
\title{Create ggplot of sonde project}
\usage{
plot_sonde(
  data,
  y_var,
  opts = list(points = TRUE, line = TRUE, files = FALSE, oow = FALSE, calcheck = FALSE,
    precip = FALSE),
  fieldform = NULL,
  calcheck = NULL,
  precip = NULL
)
}
\arguments{
\item{data}{A `data.frame` with the sonde data to plot.}

\item{y_var}{Y-variable to plot on the y-axis.}

\item{opts}{A list of options for plotting:
- points: should points be plotted?
- line: should line be plotted?
-files: should points be colored by file?
-oow: should out of water periods be plotted?
-calcheck: should cal check be plotted?
-precip: should precip data be plotted?}

\item{fieldform}{Field form from the `sondeproj`.}

\item{calcheck}{Calibration check from the `sondeproj`.}

\item{precip}{Precip data from the `sondeproj`.}
}
\value{
a `ggplot2` object
}
\description{
Allows options for plotting sonde project data for different variables, used within modules to
simplify code for plotting
}
\examples{
plot_sonde(example_data, "Temp_C")
}
