% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-interp.R
\name{prep_interp}
\alias{prep_interp}
\title{Prepare interpolation dataset}
\usage{
prep_interp(proj)
}
\arguments{
\item{proj}{A \code{sondeproj} object holding sonde data.}
}
\value{
a \code{data.frame} based on \code{proj$data} with duplicates condensed to a single value for interpolating.
}
\description{
Completes the date-time sequence to add missing rows for missing datetimes. Also creates an
interpolation dataset where any duplicates have been condensed to a single value. In the case of
two different values, they will be set to NA for the purposes of interpolation to not
interpolate with questionable data.
}
\examples{
interp_dfs <- prep_interp(example_sondeproj)
}
