% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4-quality-flags.R
\name{quality_UI}
\alias{quality_UI}
\alias{quality_server}
\title{Flag data as questionable}
\usage{
quality_UI(id)

quality_server(id, sondeproj, data_ver, y_var)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}

\item{y_var}{Y-variable to plot on the y-axis.}
}
\description{
Allows the user to mark data that looks for "weird" and can be viewed as a plotting option or
used to auto-select as an outlier.
}
\keyword{internal}
