% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-sonde-project.R
\name{save_project}
\alias{save_project}
\alias{read_project}
\title{Save version history and sonde data as a project}
\usage{
save_project(data_ver, log, prj_path)

read_project(prj_path)
}
\arguments{
\item{data_ver}{the data version object (a list)}

\item{log}{the change log object (a data.frame)}

\item{prj_path}{the path to save the project to including the file name}
}
\value{
saves data_ver and log as a .qs file
}
\description{
Combines the change history with the version history of the sonde data and saves to a
\code{.qs} file to prevent data loss and enable work over multiple days.
}
