% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sonde-csv.R
\name{read_sonde}
\alias{read_sonde}
\title{Upload raw sonde data}
\usage{
read_sonde(file, encoding = NULL, flags = TRUE, skip = NULL, tz = NULL)
}
\arguments{
\item{file}{file path to sonde data (.csv)}

\item{encoding}{the file encoding, will guess if NULL}

\item{flags}{logical, if TRUE a flag column will be added for each parameter}

\item{skip}{the number of rows to skip, assumes the first row is a header,  will guess if NULL}

\item{tz}{the time zone for the dataset. If not specified will use user timezone. See details for more information}
}
\value{
a data.frame of sonde data
}
\description{
Loads a \code{.csv} file with raw sonde data. Cleans column names, adds a \code{DateTime} column,
and optionally adds a \code{flag} column.
}
\details{
Daylights saving time can cause issues for continous datasets, so data is often collected in standard time.
If this is the case use Etc/GMT offsets. See \link[base]{OlsonNames} for all available timezones.
}
\examples{
file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-example.csv")
df <- read_sonde(file)

file <- file.path(fs::path_package("extdata", package = "SondePolishR"), "sonde-usb-example.csv")
df <- read_sonde(file)
}
