% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift-points.R
\name{shift_points}
\alias{shift_points}
\title{Correct points via an absolute shift}
\usage{
shift_points(data, par, index, shift_val = NULL)
}
\arguments{
\item{data}{a data.frame with sonde data}

\item{par}{the parameter being corrected}

\item{index}{the index values of the rows that need to be shifted}

\item{shift_val}{the value to shift the selected data by, if \code{NULL}, it will be guessed using \link[SondePolishR]{guess_shift}}
}
\value{
a data.frame with the values adjusted
}
\description{
Sometimes a continuous sonde signal will be shifted up or down by a consistent value for a period of time, this will adjust the
dataset for those values by that set amount.
}
\examples{
raw_sonde$Cond_uS_cm[1591:1600]
data <- shift_points(raw_sonde, "Cond_uS_cm", 1591:1630)
data$Cond_uS_cm[1591:1600]
}
