% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-data-module.R
\name{update_parms_UI}
\alias{update_parms_UI}
\alias{update_parms_server}
\title{Shiny module to update plotting parameters}
\usage{
update_parms_UI(id)

update_parms_server(id, sondeproj, choices_fun = NULL)
}
\arguments{
\item{id}{An ID string passed to shiny::NS(), used for namespacing UI inputs/outputs.}

\item{sondeproj}{A \code{reactiveVal} holding the current dataset.}

\item{choices_fun}{Function used to determine the parameter choices, if \code{NULL} will use the column names of the data}
}
\value{
the selected variable \code{y_var} as a reactive object
}
\description{
Used to dynamically update the choices for selecting a y variable based on the names in the data or
a custom function.
}
\keyword{internal}
