% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_1-weekly-dates-module.R
\name{weekly_range_sidebar_UI}
\alias{weekly_range_sidebar_UI}
\alias{weekly_range_buttons_UI}
\alias{weekly_range_server}
\title{View data by a specified time length}
\usage{
weekly_range_sidebar_UI(id)

weekly_range_buttons_UI(id)

weekly_range_server(id, sondeproj, period_view, dates, p_length, data_ver)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{sondeproj}{A `reactiveVal` holding the current dataset.}

\item{period_view}{Should data be viewed by period?}

\item{dates}{The date range to view the data.}

\item{p_length}{The length of the period to view.}

\item{data_ver}{A `reactiveVal` holding a number used to track when new data is added to trigger resets.}
}
\value{
a reactive of length two with the min and max dates.
}
\description{
Used to adjust plotting to show periods to better examine details.
}
\keyword{internal}
