% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1-weekly-dates-module.R
\name{weekly_range_sidebar_UI}
\alias{weekly_range_sidebar_UI}
\alias{weekly_range_buttons_UI}
\alias{weekly_range_server}
\title{Return weekly date ranges}
\usage{
weekly_range_sidebar_UI(id)

weekly_range_buttons_UI(id)

weekly_range_server(id, min_date, max_date)
}
\arguments{
\item{id}{the shiny ID of the module}

\item{min_date}{Minimum date in the `sondeproj`.}

\item{max_date}{Maximum date in the `sondeproj`.}
}
\value{
a reactive of length two with the min and max dates.
}
\description{
Used to adjust plotting to show weekly periods to better examine details.
}
\keyword{internal}
